slowly move on to use editor project for gui, renderer system refactor and abstraction for the renderers management

+++ Engine/src/core/resources/managers/BufferManager.h
@@ -15,10 +15,8 @@ public:
     virtual void destroy(uint32_t id) = 0;
     virtual std::vector<uint32_t> listIDs() const = 0;
 
-
-
     virtual uint32_t createVertexBuffer(const Vertex* vertices, int size) = 0;
-    virtual uint32_t createIndexBuffer(const uint16_t* indices, int size) = 0;
+    virtual uint32_t createIndexBuffer(const uint32_t* indices, int size) = 0;
     virtual void bind(uint32_t id) = 0;
     virtual Buffer* getBuffer(uint32_t id) = 0;
